Rename utils to helpers and move some basic service functions to helpers
